requirements.txt
fastapi==0.75.2
python-dotenv==0.20.0
requests==2.27.1
uvicorn==0.17.6
gunicorn==20.1.0